@@ -10,20 +10,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.0",
-    "@angular/cdk": "^17.2.1",
-    "@angular/common": "^17.1.0",
-    "@angular/compiler": "^17.1.0",
-    "@angular/core": "^17.1.0",
-    "@angular/elements": "^18.2.0",
-    "@angular/fire": "^17.0.1",
-    "@angular/forms": "^17.1.0",
-    "@angular/google-maps": "^17.2.1",
-    "@angular/material": "^17.2.1",
-    "@angular/platform-browser": "^17.1.0",
-    "@angular/platform-browser-dynamic": "^17.1.0",
-    "@angular/router": "^17.1.0",
-    "@angular/youtube-player": "^17.2.1",
+    "@angular/animations": "^19.2.14",
+    "@angular/cdk": "^18.2.14",
+    "@angular/common": "^19.2.14",
+    "@angular/compiler": "^19.2.14",
+    "@angular/core": "^18.2.13",
+    "@angular/elements": "^19.2.14",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^19.2.14",
+    "@angular/google-maps": "^18.2.14",
+    "@angular/material": "^18.2.14",
+    "@angular/platform-browser": "^19.2.14",
+    "@angular/platform-browser-dynamic": "^19.2.14",
+    "@angular/router": "^19.2.14",
+    "@angular/youtube-player": "^18.2.14",
     "@fireworks-js/angular": "^2.10.8",
     "@ionic/angular": "^7.7.3",
     "@ngx-translate/core": "^15.0.0",
@@ -39,16 +39,16 @@
     "ng2-charts": "^6.0.0",
     "ngx-mat-select-search": "^8.0.2",
     "ngx-stripe": "^17.1.1",
-    "ngx-toastr": "^18.0.0",
+    "ngx-toastr": "^19.0.0",
     "rxjs": "~7.8.0",
     "socket.io-client": "^4.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.3"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.2",
-    "@angular/cli": "^17.1.2",
-    "@angular/compiler-cli": "^17.1.0",
+    "@angular-devkit/build-angular": "^18.2.20",
+    "@angular/cli": "^18.2.20",
+    "@angular/compiler-cli": "^19.2.14",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
@@ -57,6 +57,6 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
     "tailwindcss": "^3.4.1",
-    "typescript": "~5.3.2"
+    "typescript": "~5.5.4"
   }
 }
