@@ -1,10 +1,5 @@
 import { Injectable } from '@angular/core';
-import {
-  HttpInterceptor,
-  HttpRequest,
-  HttpHandler,
-  HttpParams
-} from '@angular/common/http';
+import { HttpInterceptor, HttpRequest, HttpHandler, HttpParams } from '@angular/common/http';
 import { take, exhaustMap } from 'rxjs/operators';
 
 import { AuthService } from './auth.service';
